Add OperationsForm composable to upsert transaction in database; Transaction class contains all necessary information to be used in Composable; update kotlin version 2.0.10; store CategoryTypes by its ordinal in Enum class; remove trailing content in Transactions Day summary in Operations screen to be done other way;

+++ app/src/main/java/com/qosquo/wallet/data/db/entity/CategoriesDbEntity.kt
@@ -11,7 +11,7 @@ import androidx.room.PrimaryKey
 data class CategoriesDbEntity(
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "category_name") val categoryName: String,
-    val type: Short,
+    val type: Int,
     val goal: Float,
     @ColumnInfo(name = "category_icon_id") @DrawableRes val categoryIconId: Int,
     @ColumnInfo(name = "color_hex") val colorHex: String
